WIP add basic i18n
This is a work in progress. I just wanted to play around with the Rails
i18n to see what it feels like.

Ideally, we would rely on the Accept-Language HTTP header. But, as a
quick check I decided to check for a query string parameter.

More info on using the Accept-Language HTTP header can be found at:

  http://guides.rubyonrails.org/i18n.html#setting-the-locale-from-the-client-supplied-information

Change-Id: I314331dd373dfb6309f50633dd98320b040b8b96